fix: extract shared formatReviewDate helper and fix UTC timezone issue in review dates
- Add formatReviewDate() to date_helper.ts as an exported reusable helper
  that formats YYYY-MM-DD strings via Parse_EN_US + FormatIntl, avoiding
  the UTC off-by-one day issue that new Date('YYYY-MM-DD') causes in
  non-UTC timezones
- Update ReviewsPageView.ts to import formatReviewDate from date_helper
  and use it for review createdAt dates (previously used
  toLocaleDateString(undefined, ...) which was locale-unstable)
- Update ReviewsPageView test to compute the expected date value via
  formatReviewDate instead of a hardcoded locale-dependent string
